Extract some utils from controller
