+++ src/controller.js
@@ -8,124 +8,13 @@ const eventStream = require('event-stream');
 const mongoose = require('mongoose');
 const semver = require('semver');
 
-// A filter function for checking a given value is defined and not null.
-function exists(o) {
-  return o !== undefined && o !== null;
-}
-// Handle variable number of arguments
-function last(skip, names, values) {
-  const r = {};
-  let position = names.length;
-  const count = values.filter(exists).length - skip;
-  if (count < 1) throw RestError.Misconfigured('Too few arguments.');
+const {
+  isPositiveInteger,
+  getAsInt,
+  isNonNegativeInteger
+} = require('./utils/predicates-and-accessors');
+const {defineRoutes} = require('./utils/routing');
 
-  names.forEach(function(name) {
-    const index = skip + count - position;
-    position--;
-    if (index >= skip) r[name] = values[index];
-  });
-
-  return r;
-}
-// Returns `true` if the given stirng is a recognized HTTP method.
-function isRecognizedMethod(s) {
-  return !!/^all|head|get|put|post|delete$/.exec(s);
-}
-// Parse middleware into an array of middleware definitions for each endpoint and method
-function factor(options) {
-  const factored = [];
-  let methodString = options.methods;
-
-  if (methodString) methodString = methodString.toLowerCase();
-
-  if (!methodString || methodString === '*') methodString = 'all';
-  const methods = methodString.split(/\s+/);
-
-  methods.forEach(function(method) {
-    if (!isRecognizedMethod(method))
-      throw RestError.Misconfigured('Unrecognized HTTP method: "%s"', method);
-  });
-
-  if (!options.stage) throw RestError.Misconfigured('The middleware stage was not provided');
-  if (options.endpoint && options.endpoint !== 'instance' && options.endpoint !== 'collection') {
-    throw RestError.Misconfigured(
-      'End-point type must be either "instance" or "collection," not "%s"',
-      options.endpoint
-    );
-  }
-  // Middleware function or array
-  if (!Array.isArray(options.middleware) && typeof options.middleware !== 'function') {
-    throw RestError.Misconfigured('Middleware must be an array or function');
-  }
-  // Check endpoint is valid
-  if (
-    options.endpoint !== undefined &&
-    options.endpoint !== 'instance' &&
-    options.endpoint !== 'collection'
-  ) {
-    throw RestError.Misconfigured(
-      'End-point type must be either "instance" or "collection," not "%s"',
-      options.endpoint
-    );
-  }
-  // Add definitions for one or both endpoints, for each HTTP method.
-  methods.forEach(function(method) {
-    if (options.endpoint !== 'collection')
-      factored.push({
-        stage: options.stage,
-        endpoint: 'instance',
-        method,
-        middleware: options.middleware
-      });
-    if (options.endpoint !== 'instance')
-      factored.push({
-        stage: options.stage,
-        endpoint: 'collection',
-        method,
-        middleware: options.middleware
-      });
-  });
-
-  return factored;
-}
-
-// Expands route definitions based on generalized arguments.
-const defineRoutes = function(stage, params) {
-  const argumentsArray = Array.prototype.slice.call(params);
-
-  const options = last(0, ['endpoint', 'methods', 'middleware'], argumentsArray);
-  options.stage = stage;
-
-  return factor(options);
-};
-
-// query.options methods
-
-function isDefinedAndNotNull(n) {
-  if (n === null) return false;
-  if (n === undefined) return false;
-  return true;
-}
-
-function isPositiveInteger(n) {
-  if (!isDefinedAndNotNull(n)) return false;
-  n = Number(n);
-  if (n < 1) return false;
-  return n === Math.ceil(n);
-}
-
-function getAsInt(n) {
-  return Math.ceil(Number(n));
-}
-
-function isNonNegativeInteger(n) {
-  if (!isDefinedAndNotNull(n)) return false;
-  n = Number(n);
-  if (n < 0) return false;
-  return n === Math.ceil(n);
-}
-
-// __Module Definition__
 const Controller = deco(function(model, protect) {
   const controller = this;
   const initial = express.Router();
@@ -137,7 +26,6 @@ const Controller = deco(function(model, protect) {
   };
 
   // §CONFIGURE
-
   if (typeof model !== 'string' && (!model || !model.schema)) {
     throw RestError.Misconfigured('You must pass in a model or model name');
   }
